### Motivation / Background

I came across this code while using memory_profiler on my test suite.

The current implementation creates 2 Arrays and 1 Set. Looking more closely, it seemed like it should be possible to do all the operations starting with an initial Set, and keep adding to that rather than creating extra copies of Arrays.

I ran some benchmarks comparing the original implementation and this new one, and I wasn't surprised to see less allocated memory, but I was shocked by how much faster faster it was... some 508x!

```
Comparison:
          refactored:       7704 allocated
            original:       8352 allocated - 1.08x more
Warming up --------------------------------------
            original     2.293k i/100ms
          refactored     1.189M i/100ms
Calculating -------------------------------------
            original     23.489k (± 1.0%) i/s -    119.236k in   5.076701s
          refactored     11.951M (± 0.6%) i/s -     60.640M in   5.074167s

Comparison:
            original:    23489.3 i/s
          refactored: 11951077.0 i/s - 508.79x  faster
```

### Detail

This Pull Request changes the implementation of AbstractController::Base#action_methods to be more memory efficient and faster.

### Additional information

I used this for benchmarking:

```ruby
# frozen_string_literal: true
require 'action_controller'
require 'benchmark/ips'
require 'benchmark/memory'

module Actions
  def foo
  end

  def bar
  end

  def baz
  end
end

class Current < ActionController::Base
  include Actions
  def action_methods
    @action_methods ||= begin
      # All public instance methods of this class, including ancestors
      methods = (public_instance_methods(true) -
        # Except for public instance methods of Base and its ancestors
        internal_methods +
        # Be sure to include shadowed public instance methods of this class
        public_instance_methods(false))

      methods.map!(&:to_s)

      methods.to_set
    end
  end
end

class Refactored2 < ActionController::Base
  include Actions

  class << self
      def action_methods
        @action_methods ||= begin
                              Set.new(public_instance_methods(true))
                                .subtract(internal_methods)
                                .merge(public_instance_methods(false))
                                .map!(&:to_s)
        end
      end
  end
end

Benchmark.memory do |x|
  x.report("original") do
    Current.clear_action_methods!
    Current.action_methods
  end

  x.report("refactored") do
    Current.clear_action_methods!
    Refactored2.action_methods
  end

  x.compare!
end

Benchmark.ips do |x|
  x.report("original") do
    Current.clear_action_methods!
    Current.action_methods
  end

  x.report("refactored") do
    Current.clear_action_methods!
    Refactored2.action_methods
  end

  x.compare! order: :baseline
end
```

Results are included above.

I tried at least one other implementation, where I started with a new Set, but the final result was marginally faster:

```ruby
                              methods = Set.new
                              methods.merge(public_instance_methods(true))
                              methods.subtract(internal_methods)
                              methods.merge(public_instance_methods(false))
                              methods.map!(&:to_s)

```

### Checklist

Before submitting the PR make sure the following are checked:

* [x] This Pull Request is related to one change. Changes that are unrelated should be opened in separate PRs.
* [x] Commit message has a detailed description of what changed and why. If this PR fixes a related issue include it in the commit message. Ex: `[Fix #issue-number]`
* [x] Tests are added or updated if you fix a bug or add a feature.
* [x] CHANGELOG files are updated for the changed libraries if there is a behavior change or additional feature. Minor bug fixes and documentation changes should not be included.