@rafaelfranca 2023-08-23
Should not

```
  x.report("refactored") do
    Current.clear_action_methods!
    Refactored2.action_methods
  end
```

be

```
  x.report("refactored") do
    Refactored2.clear_action_methods!
    Refactored2.action_methods
  end
```

Hitting the memoized value would explain why it is 500x faster.

If you change that, you get

```
Comparison:
            original:    20211.5 i/s
          refactored:    11931.6 i/s - 1.69x  (± 0.00) slower
```