@technicalpickles 2023-08-23
Oof, you are totally right 😮‍💨 

There is still less memory allocations, but that isn't worth the performance hit. I'll play around with it a bit more and re-open if I can find something