@tenderlove 2023-08-23
`action_methods` is a class method that caches its results, isn't it?  Maybe we could make it faster (or use less memory), but since it should only ever perform its calculation once I'm not sure how it would be a bottleneck.  Maybe there's a way to pre-heat caches like this before running benchmarks (or there's a bug that it isn't being cached?)